requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.3.0
pyocr==0.4.1
pytesseract==0.1.6
redis==2.10.5
six==1.10.0
Werkzeug==0.11.10
